pyproject.toml
[project]
name = "backend"
version = "0.0.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncpg>=0.30.0",
    "authlib>=1.6.5",
    "fastapi[standard]>=0.122.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "jx>=0.5.1",
    "markdown>=3.10",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pygments>=2.19.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "pytest>=9.0.1",
    "ruff>=0.14.6",
    "ty>=0.0.1a28",
]
